requirements.txt
annotated-types==0.5.0 ; python_version >= "3.10" and python_version < "4.0"
anyio==3.7.1 ; python_version >= "3.10" and python_version < "4.0"
argcomplete==3.1.1 ; python_version >= "3.10" and python_version < "4.0"
attrs==23.1.0 ; python_version >= "3.10" and python_version < "4.0"
black==23.7.0 ; python_version >= "3.10" and python_version < "4.0"
cachetools==5.3.1 ; python_version >= "3.10" and python_version < "4.0"
certifi==2023.7.22 ; python_version >= "3.10" and python_version < "4.0"
chardet==5.2.0 ; python_version >= "3.10" and python_version < "4.0"
charset-normalizer==3.2.0 ; python_version >= "3.10" and python_version < "4.0"
click==8.1.6 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
datamodel-code-generator[http]==0.21.4 ; python_version >= "3.10" and python_version < "4.0"
dnspython==2.4.2 ; python_version >= "3.10" and python_version < "4.0"
email-validator==2.0.0.post2 ; python_version >= "3.10" and python_version < "4.0"
exceptiongroup==1.1.2 ; python_version >= "3.10" and python_version < "3.11"
genson==1.2.2 ; python_version >= "3.10" and python_version < "4.0"
google-api-core==2.11.1 ; python_version >= "3.10" and python_version < "4.0"
google-auth==2.17.3 ; python_version >= "3.10" and python_version < "4.0"
google-cloud-core==2.3.3 ; python_version >= "3.10" and python_version < "4.0"
google-cloud-storage==2.10.0 ; python_version >= "3.10" and python_version < "4.0"
google-crc32c==1.5.0 ; python_version >= "3.10" and python_version < "4.0"
google-resumable-media==2.5.0 ; python_version >= "3.10" and python_version < "4.0"
googleapis-common-protos==1.60.0 ; python_version >= "3.10" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.10" and python_version < "4.0"
httpcore==0.17.3 ; python_version >= "3.10" and python_version < "4.0"
httpx==0.24.1 ; python_version >= "3.10" and python_version < "4.0"
idna==3.4 ; python_version >= "3.10" and python_version < "4.0"
inflect==5.6.2 ; python_version >= "3.10" and python_version < "4.0"
isort==5.12.0 ; python_version >= "3.10" and python_version < "4.0"
jinja2==3.1.2 ; python_version >= "3.10" and python_version < "4.0"
jsonschema-spec==0.1.6 ; python_version >= "3.10" and python_version < "4.0"
jsonschema==4.17.3 ; python_version >= "3.10" and python_version < "4.0"
lazy-object-proxy==1.9.0 ; python_version >= "3.10" and python_version < "4.0"
markupsafe==2.1.3 ; python_version >= "3.10" and python_version < "4.0"
mypy-extensions==1.0.0 ; python_version >= "3.10" and python_version < "4.0"
openapi-schema-validator==0.4.4 ; python_version >= "3.10" and python_version < "4.0"
openapi-spec-validator==0.5.7 ; python_version >= "3.10" and python_version < "4.0"
packaging==23.1 ; python_version >= "3.10" and python_version < "4.0"
pathable==0.4.3 ; python_version >= "3.10" and python_version < "4.0"
pathspec==0.11.2 ; python_version >= "3.10" and python_version < "4.0"
platformdirs==3.10.0 ; python_version >= "3.10" and python_version < "4.0"
prance==23.6.21.0 ; python_version >= "3.10" and python_version < "4.0"
protobuf==4.24.0 ; python_version >= "3.10" and python_version < "4.0"
pyasn1-modules==0.3.0 ; python_version >= "3.10" and python_version < "4.0"
pyasn1==0.5.0 ; python_version >= "3.10" and python_version < "4.0"
pydantic-core==2.4.0 ; python_version >= "3.10" and python_version < "4.0"
pydantic==2.1.1 ; python_version >= "3.10" and python_version < "4.0"
pydantic[email]==2.1.1 ; python_version >= "3.10" and python_version < "4.0"
pyrsistent==0.19.3 ; python_version >= "3.10" and python_version < "4.0"
pysnooper==1.2.0 ; python_version >= "3.10" and python_version < "4.0"
python-dotenv==1.0.0 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.10" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.10" and python_version < "4.0"
rfc3339-validator==0.1.4 ; python_version >= "3.10" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.10" and python_version < "4"
ruamel-yaml-clib==0.2.7 ; platform_python_implementation == "CPython" and python_version < "3.12" and python_version >= "3.10"
ruamel-yaml==0.17.32 ; python_version >= "3.10" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.10" and python_version < "4.0"
toml==0.10.2 ; python_version >= "3.10" and python_version < "4.0"
tomli==2.0.1 ; python_version >= "3.10" and python_version < "3.11"
typing-extensions==4.7.1 ; python_version >= "3.10" and python_version < "4.0"
urllib3==2.0.4 ; python_version >= "3.10" and python_version < "4.0"
